requirements/base.txt
Django==5.0.8
psycopg2-binary==2.9.10
python-dotenv==1.0.1
whitenoise==6.2.0
django-extensions==3.2.1
pillow==10.4.0




djangorestframework==3.15.2
django-filter==24.3
markdown==3.6
djangorestframework-simplejwt~=5.3.1
django-cors-headers~=4.4
drf-spectacular==0.27.0



model-bakery==1.19.1
# Faker==15.1.1